+++ src/refactorings/extract/extract-variable/extract-variable.extractable.test.ts
@@ -1,5 +1,5 @@
-import { Code } from "../../../editor/editor";
 import { InMemoryEditor } from "../../../editor/adapters/in-memory-editor";
+import { Code } from "../../../editor/editor";
 import { testEach } from "../../../tests-helpers";
 
 import { extractVariable } from "./extract-variable";
@@ -542,6 +542,29 @@ const extracted = a + 1;
 console.log(a);
 console.log(extracted);
 console.log(extracted);`
+      },
+      {
+        description: "expression bound to a variable, nested in statements",
+        code: `function updateQuality() {
+  for (var i = 0; i < items.length; i++) {
+    if ([start]items[i][end].name != "SULFURAS") {
+      items[i].sellIn = items[i].sellIn - 1;
+    }
+  }
+
+  return items;
+}`,
+        // The declared variable isn't indented, but at least it doesn't mess up with the next line (if statement).
+        expected: `function updateQuality() {
+  for (var i = 0; i < items.length; i++) {
+const extracted = items[i];
+    if (extracted.name != "SULFURAS") {
+      extracted.sellIn = extracted.sellIn - 1;
+    }
+  }
+
+  return items;
+}`
       }
     ],
     async ({ code, expected }) => {